+++ app/controllers/emails_controller.rb
@@ -4,18 +4,7 @@
 class EmailsController < ApplicationController
   include ::Emailable
 
-  before_action :login_required, except: [
-    :ask_webmaster_question
-  ]
-
-  def ask_webmaster_question
-    @email = params.dig(:user, :email)
-    @content = params.dig(:question, :content)
-    @email_error = false
-    return create_webmaster_question if request.method == "POST"
-
-    @email = @user.email if @user
-  end
+  before_action :login_required
 
   def merge_request
     return unless (@model = validate_merge_model!(params[:type]))
@@ -53,30 +42,6 @@ class EmailsController < ApplicationController
   ##########
 
   private
-
-  def create_webmaster_question
-    if @email.blank? || @email.index("@").nil?
-      flash_error(:runtime_ask_webmaster_need_address.t)
-      @email_error = true
-    elsif @content.blank?
-      flash_error(:runtime_ask_webmaster_need_content.t)
-    elsif non_user_potential_spam?
-      flash_error(:runtime_ask_webmaster_antispam.t)
-    else
-      QueuedEmail::Webmaster.create_email(sender_email: @email,
-                                          content: @content)
-      flash_notice(:runtime_ask_webmaster_success.t)
-      redirect_to("/")
-    end
-  end
-
-  def non_user_potential_spam?
-    !@user && (
-      /https?:/.match?(@content) ||
-      %r{<[/a-zA-Z]+>}.match?(@content) ||
-      @content.exclude?(" ")
-    )
-  end
 
   def validate_merge_model!(val)
     case val
