+++ app/views/controllers/admin/emails/webmaster_questions/new.erb
@@ -1,0 +1,6 @@
+<%
+add_page_title(:ask_webmaster_title.t)
+%>
+
+<%= render(partial: "admin/emails/webmaster_questions/form",
+           locals: { local: true }) %>
