+++ app/views/controllers/application/sidebar/_info.html.erb
@@ -9,9 +9,9 @@ active_link_to(:app_donate.t, support_donate_path,
                class: classes[:indent], id: "nav_donate_link"),
 active_link_to(:app_how_to_help.t, info_how_to_help_path,
                class: classes[:indent], id: "nav_how_to_help_link"),
-active_link_to(:app_report_a_bug.t, emails_ask_webmaster_question_path,
+active_link_to(:app_report_a_bug.t, new_admin_emails_webmaster_questions_path,
                class: classes[:indent], id: "nav_bug_report_link"),
-active_link_to(:app_send_a_comment.t, emails_ask_webmaster_question_path,
+active_link_to(:app_send_a_comment.t, new_admin_emails_webmaster_questions_path,
                class: classes[:indent], id: "nav_ask_webmaster_link"),
 active_link_to(:app_contributors.t, contributors_path,
                class: classes[:indent], id: "nav_contributors_link"),
